Addes Roll History
Added Roll History.

After a roll (to prevent bombarding the history record if the user spam clicks individual dice, a roll is concidered... "rolled" when all the dice inside the tray stop rolling. This might lead to abuse if the player keeps the dice rolling by clicking on another die while others are rolling, but this is a feature beyon this scope) a record gets inserted in a log that is the displayed on a drop down list.